+++ forgero-core/src/test/java/com/sigmundgranaas/forgeroforge/testutil/Tools.java
@@ -1,14 +1,14 @@
 package com.sigmundgranaas.forgeroforge.testutil;
 
 import com.sigmundgranaas.forgero.core.property.attribute.Category;
-import com.sigmundgranaas.forgero.core.state.Composite;
 import com.sigmundgranaas.forgero.core.state.Ingredient;
-import com.sigmundgranaas.forgero.core.state.slot.EmptySlot;
+import com.sigmundgranaas.forgero.core.state.composite.Construct;
+import com.sigmundgranaas.forgero.core.state.upgrade.slot.EmptySlot;
 
 import java.util.Set;
 
 public class Tools {
-    public static Composite IRON_PICKAXE = Composite.builder()
+    public static Construct IRON_PICKAXE = Construct.builder()
             .addIngredient(Ingredient.of(ToolParts.HANDLE))
             .addIngredient(Ingredient.of(ToolParts.PICKAXE_HEAD))
             .addUpgrade(new EmptySlot(0, Types.BINDING, "", Set.of(Category.UTILITY)))