+++ forgero-fabric/src/test/java/com/sigmundgranaas/forgerofabric/nbt/v2/StateNbtConversionTest.java
@@ -3,8 +3,8 @@ package com.sigmundgranaas.forgerofabric.nbt.v2;
 import com.sigmundgranaas.forgero.core.ForgeroStateRegistry;
 import com.sigmundgranaas.forgero.core.property.AttributeType;
 import com.sigmundgranaas.forgero.core.resource.PipelineBuilder;
-import com.sigmundgranaas.forgero.core.state.Composite;
 import com.sigmundgranaas.forgero.core.state.State;
+import com.sigmundgranaas.forgero.core.state.composite.Construct;
 import com.sigmundgranaas.forgero.fabric.resources.FabricPackFinder;
 import com.sigmundgranaas.forgero.minecraft.common.item.nbt.v2.CompositeEncoder;
 import com.sigmundgranaas.forgero.minecraft.common.item.nbt.v2.CompositeParser;
@@ -46,7 +46,7 @@ public class StateNbtConversionTest {
     @Test
     void encodeCompoundParseCompoundWithIngredients() {
         NbtCompound compound = encoder.encode(Tools.IRON_PICKAXE);
-        var pickaxe = parser.parse(compound).map(Composite.class::cast).orElseThrow();
+        var pickaxe = parser.parse(compound).map(Construct.class::cast).orElseThrow();
         Assertions.assertEquals(2, pickaxe.ingredients().size());
         Assertions.assertEquals("oak-handle", pickaxe.ingredients().get(0).name());
         Assertions.assertEquals("iron-pickaxe_head", pickaxe.ingredients().get(1).name());
@@ -55,7 +55,7 @@ public class StateNbtConversionTest {
     @Test
     void encodeCompoundParseCompoundWithUpgrades() {
         NbtCompound compound = encoder.encode(Tools.IRON_PICKAXE.upgrade(Upgrades.BINDING));
-        var pickaxe = StateParser.STATE_PARSER.parse(compound).map(Composite.class::cast).orElseThrow();
+        var pickaxe = StateParser.STATE_PARSER.parse(compound).map(Construct.class::cast).orElseThrow();
         //Assertions.assertEquals(0, pickaxe.upgrades().size());
         //Assertions.assertEquals("oak-binding", pickaxe.upgrades().get(0).name());
     }
@@ -63,7 +63,7 @@ public class StateNbtConversionTest {
     @Test
     void encodeCompoundParseCompoundWithProperties() {
         NbtCompound compound = encoder.encode(Tools.IRON_PICKAXE.upgrade(Upgrades.BINDING));
-        var pickaxe = parser.parse(compound).map(Composite.class::cast).orElseThrow();
+        var pickaxe = parser.parse(compound).map(Construct.class::cast).orElseThrow();
         Assertions.assertEquals(0, pickaxe.stream().applyAttribute(AttributeType.ATTACK_DAMAGE));
         Assertions.assertEquals(0, pickaxe.stream().applyAttribute(AttributeType.DURABILITY));
     }